gpt-5-5: decompose TimelineSection test monolith into focused suites (#259)
Split TimelineSection.test.tsx into rendering, scroll, typewriter, and ordering
files with shared vi.hoisted mock setup. Replaced readFileSync source assertions
with behavioral useTimelineScroll contract test; deduped duplicate #237 block.

Files: TimelineSection.{rendering,scroll,typewriter,ordering}.test.tsx,
TimelineSection.test-setup.ts; deleted monolith test file.
